add native segwit support for ledger message signing
