+++ ledger/index.ts
@@ -1,6 +1,8 @@
 export * from './btc';
 export * from './psbt';
 export * from './stx';
+export * from './stx';
+export * from './btcMessageSigning';
 export * from './types';
 
 export { getMasterFingerPrint } from './helper';
